Fixed all the compile time errors
Add projects of GreeStore and Repository thus resolving the compile time
errors
Added dependency on Entity Framework
The project Repository still needs to be implemented

+++ Main/Solution/Site/GetEmissionRequest.cs
@@ -10,9 +10,9 @@ namespace GreenBoxService
         public string VehicleSubTypeId { get; set; }
 
         [DataMember]
-        public decimal[] Distance { get; set; }
+        public double[] Distance { get; set; }
 
         [DataMember]
-        public string MapXml{get;set;}
+        public string MapXml { get; set; }
     }
 }
